+++ umami/models/model_umami.py
@@ -172,7 +172,7 @@ def create_umami_model(
     return umami, nn_structure["epochs"], init_epoch
 
 
-def umami_tagger(args, train_config, preprocess_config):
+def umami_tagger(args, train_config):
     """Training handling of UMAMI tagger.
 
     Parameters
@@ -181,8 +181,6 @@ def umami_tagger(args, train_config, preprocess_config):
         Arguments from command line parser
     train_config : object
         training configuration
-    preprocess_config : object
-        preprocessing configuration
 
     Raises
     ------
@@ -215,7 +213,6 @@ def umami_tagger(args, train_config, preprocess_config):
     if n_jets_val > 0:
         val_data_dict = utt.load_validation_data_umami(
             train_config=train_config,
-            preprocess_config=preprocess_config,
             n_jets=n_jets_val,
             convert_to_tensor=True,
         )